Move select options view helpers into their own module
At some point the course choice screens should be refactored into
ActionView::Components, removing the need for these helpers. For now,
organise them a little better.
